feat: `get()` returns Readonly<T>, and deep-equal support added.
`get()` is now slightly more safe as typescript will show errors if you
attempt to modify root properties directly. It is not perfect since it
is not recursively readonly, but that is ok, we don't want to
recursively infer since objects could be huge.

deepEqual support was added by including a very fast built-in
comparator-function. Also allows the user to specify a custom comparator-fn.